improvement(tables): cleanup — extract components, stabilize callbacks, fix ref sync
